+++ test/end-to-end/commands/local-clean.sh
@@ -3,6 +3,7 @@
 pm2 flush
 pm2 delete controller
 pm2 delete relay-domapic-module
+pm2 delete foo-service
 
 ./test/end-to-end/commands/clean-db.sh