+++ test/end-to-end/commands/start-controller.sh
@@ -1,5 +1,5 @@
 #!/usr/bin/env bash
-cd test/end-to-end/fixtures/${fixture}
+cd test/end-to-end/fixtures/controller
 rm -rf ../../../../${domapic_path}/.domapic/controller/logs
 npm i
 npm run domapic-controller start controller -- --hostName=${controller_host_name} --path=../../../../${domapic_path} --db=${db_uri} --save --authDisabled