+++ test/end-to-end/specs/console-ability-changed.specs.js
@@ -20,7 +20,7 @@ test.describe('when connection with controller was successful', function () {
   })
 
   test.it('console module should be registered in controller', () => {
-    return controllerConnection.request('/modules')
+    return controllerConnection.request('/services')
       .then(response => {
         const service = response.body.find(service => service.name === 'console')
         serviceId = service._id
@@ -39,7 +39,7 @@ test.describe('when connection with controller was successful', function () {
         const ability = response.body.find(ability => ability.name === 'stdout')
         return Promise.all([
           test.expect(response.body.length).to.equal(1),
-          test.expect(ability._module).to.equal(serviceId),
+          test.expect(ability._service).to.equal(serviceId),
           test.expect(ability._user).to.equal(serviceUserId),
           test.expect(ability.event).to.equal(false),
           test.expect(ability.action).to.equal(true),