+++ test/end-to-end/specs/service-connection.specs.js
@@ -5,10 +5,10 @@ const testUtils = require('narval/utils')
 const utils = require('./utils')
 
 test.describe('when service starts and connection options are provided', function () {
-  this.timeout(10000)
+  this.timeout(20000)
 
   test.it('should have connected to the controller', () => {
-    return utils.waitOnestimatedStartTime(5000)
+    return utils.waitOnestimatedStartTime(10000)
       .then(() => {
         return testUtils.logs.combined('service')
           .then((log) => {